refactor : 옵션 비즈니스 로직 ProductService로 이동
